+++ database/migrations/2025_03_19_043328_create_laundry_items_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id('id_laundry_item')->index();
             $table->unsignedBigInteger('id_branch_laundry_item')->nullable(); // Foreign key ke branch
             $table->string('name_laundry_item');
-            $table->bigInteger('price_laundry_item');
+            $table->integer('price_laundry_item');
             $table->time('time_laundry_item')->nullable();
             $table->text('description_laundry_item')->nullable();
             $table->enum('is_active_laundry_item', ['active', 'inactive'])->default('active');
